Separate test with new file
